+++ core/class/jazpar.class.php
@@ -722,19 +722,21 @@ $postfields = "javax.faces.partial.ajax=true&javax.faces.source=_eConsoconsoDeta
       }
       $replace['#default_unit#'] = $this->getConfiguration('defaultUnit', 'kwh');
       
-      if ($template == "jazpar2") {
+      if ($template != "jazpar") {
           $cmd = $this->getCmd(null, 'localavg');
           $min = 0;
           $max = 0;
           $avg = 0;
           $month = "?";
+          $year = "?";
           $value = "?";
           $padding = 45;
           if (is_object($cmd)) {
             $avg = round($cmd->execCmd(), 0);
             if ($avg > 0) {
                 $dateCompare = $cmd->getCollectDate();
-                $month = strftime("%B %Y", strtotime($dateCompare));
+                $month = date_fr(date('F', strtotime($dateCompare)));
+                $year = strftime("%Y", strtotime($dateCompare));
                 $cmdMonth =  $this->getCmd(null, 'consom');
                 $cmdHistory = history::byCmdIdDatetime($cmdMonth->getId(), $dateCompare);
                 if (is_object($cmdHistory)) {
@@ -767,7 +769,8 @@ $postfields = "javax.faces.partial.ajax=true&javax.faces.source=_eConsoconsoDeta
                 }
             }
           }
-          $replace['#past_month#'] = $month;
+          $replace['#past_month#'] = __($month,__FILE__);
+          $replace['#past_year#'] = $year;
           $replace['#past_month_conso#'] = $value;
           $replace['#cursor_compare#'] = $padding;
       }