+++ docs/fr_FR/changelog.md
@@ -4,6 +4,10 @@
 >
 >Pour rappel s'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
+# 08/01/2021
+- Nouveau template de widget 
+- Meilleure i18n
+
 # 01/01/2021
 - Icône flêche pour comparaison
 - Niveaux de logs modifiés
@@ -11,7 +15,6 @@
 # 30/12/2020
 - Nouvelles données récupérées du site GRDF : consommations minimum maximum et moyenne des foyers similaires.
 - Nouveau template de widget pour visualiser la comparaison de votre consommation avec des foyers similaires.
-- New widget's template available to visulalize the comparison with similar profiles
 
 # 01/12/2020
 - Gestion des periodes nulles.